Computer Science 1 Rough Draft and Final Project
For the end of the Computer Science 1 class that I took, I had to come up of a project that involved GUIs, Inheritance, Polymorphism, and all other concepts learned during the course.  This project had its challenges as it was difficult  to automate the score and keep track of all the players and the scores for every correct or incorrect answers. The player can go through the menu to print out score on GUI or into a text file. This was one of the projects in Java that I was proud off and I have included the java and class files for the Rough Draft along with the code of the final project.